Full commit message for to the previous commit
Temporary workaround to fix DW_AT_stmt_list in CU generation.

I dont understand the issue here, we need a relocation here (relocation
symbol, it's index) but I dont know where to get it. For some reason
any number seems to work now. This is wrong but seems to be fixing
way too many things. Commit this for now to enable further testing
of other stuff.