Add a gross hack, fixing the FDE advance_load4 encoding in cfdwarf.
CG is giving us symbol indexes that we have to reloc_pair.
Unfortunately this is not doable with plain libdwarf interface
so we invent the very same hack that pathcc did to sgi libdwarf.

Ie. we encode the two symbol indexes into 32bit number. And when
we are writing out the FDE instructions we search for DW_CFE_advanceload4
and instead of the symbol indexes we emit the relocation.

We should really switch to .cfi_* and remove this atrocity.